Restructure packages and files; update Makefile build
