+++ cmd/dnsbench.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/askcom/dnsbench/cmd"
+import "github.com/askcom/dnsbench/pkg/cmd"
 
 func main() {
 	cmd.Execute()
